fix(frontend): satisfy org stack repo shape checks
