+++ scripts/max_lines_allowlist.txt
@@ -29,7 +29,6 @@ anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1880 existing SDK oversi
 anyharness/sdk/src/reducer/transcript.ts 1575 existing SDK oversized file
 apps/desktop/src-tauri/src/commands/keychain.rs 760 existing desktop native oversized file
 apps/desktop/src-tauri/src/commands/ssh_tunnel.rs 729 existing desktop native oversized file
-apps/desktop/src/components/workspace/chat/input/RuntimePressureIndicator.tsx 760 runtime pressure indicator with modal state and status rendering
 apps/desktop/src/lib/domain/workspaces/cloud/logical-workspaces.test.ts 662 existing desktop oversized test file
 apps/desktop/src/lib/domain/workspaces/sidebar/sidebar-indicators.test.ts 632 existing desktop oversized test file
 apps/desktop/src/lib/workflows/mcp/connector-persistence.test.ts 704 existing desktop oversized test file
